--some commit msg-- connection string added since its local proj

+++ backend/db/tests/dbo/User Defined Types/Time_UDT.sql
@@ -1,3 +1,3 @@
 ﻿CREATE TYPE [dbo].[Time_UDT]
-    FROM INT NULL;
+    FROM BIGINT NULL;
